add async job

+++ src/main/java/com/example/demo/Service/impl/StorageServiceImpl.java
@@ -1,7 +1,9 @@
 package com.example.demo.Service.impl;
 
+import com.example.demo.Bean.Task;
 import com.example.demo.Config.StorageDirConfig;
 import com.example.demo.Service.StorageService;
+import com.example.demo.Service.TaskService;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
@@ -9,19 +11,27 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.PrintWriter;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
+import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
+/**
+ * 用于处理上传存储相关的服务，自动按文件类别分类存储
+ */
 @Service
 @Slf4j
 public class StorageServiceImpl extends HttpServlet implements StorageService  {
@@ -31,28 +41,53 @@ public class StorageServiceImpl extends HttpServlet implements StorageService  {
     @Autowired
     StorageDirConfig storageDirConfig;
 
+    @Autowired
+    TaskService taskService;
+
     @Override
-    public String upload(MultipartFile[] files, String savePath) {
-
-        StringBuffer message = new StringBuffer();
-
-        for (MultipartFile file : files) {
-            String fileName = file.getOriginalFilename();
-            try {
-                Path pathToSave = getTargetLocation(fileName, savePath);
-                log.info("get pathToSave: {}", pathToSave);
-
-                File destFile = new File(String.valueOf(pathToSave));
-                if (destFile.exists()) {
-                    destFile.delete();
-                }
-                Files.copy(file.getInputStream(), pathToSave);
-                message.append("Upload file success : " + destFile.getAbsolutePath()).append("<br>");
-            } catch (IOException e) {
-                message.append("Upload file failed : " + e).append("<br>");
-            }
+    public void upload(MultipartFile[] files, String savePath, String taskId) {
+        // 从lambda表达式引用的本地变量必须是最终变量或实际上的最终变量
+        // 定义单元素数组绕过编译器限制
+        int[] i = {0};
+        long start = System.currentTimeMillis();
+        long cnt = files.length;
+        if (cnt == 0) {
+            return;
         }
-        return  message.toString();
+        taskService.updateTask(taskId, "upload", String.valueOf(i) + "/" + String.valueOf(cnt), "running");
+        List<CompletableFuture<Void>> uploadFutures = Stream.of(files)
+                .map(file -> {
+                    final String fileName = file.getOriginalFilename();
+                    Path pathToSave = getTargetLocation(fileName, savePath);
+                    return CompletableFuture.runAsync(() -> {
+                        try (InputStream inputStream = file.getInputStream()) {
+                            File destFile = new File(String.valueOf(pathToSave));
+                            if (destFile.exists()) {
+                                destFile.delete();
+                            }
+                            Files.copy(inputStream, pathToSave);
+                            log.info("Upload file success: {}", pathToSave);
+                            synchronized (this) {
+                                i[0]++;
+                                if (i[0] % 10 == 0) {
+                                    taskService.updateTask(taskId, "upload", String.valueOf(i[0]) + "/" + String.valueOf(cnt), "running");
+                                }
+                            }
+                        } catch (IOException e) {
+                            log.error("Upload file failed: {}", pathToSave, e);
+                        }
+                    });
+                })
+                .collect(Collectors.toList());
+
+        CompletableFuture<Void> allUploadsDone = CompletableFuture.allOf(uploadFutures.toArray(new CompletableFuture[0]));
+
+        allUploadsDone.thenRun(() -> {
+            int uploadedFilesCount = i[0];
+            taskService.updateTask(taskId, "upload", String.valueOf(uploadedFilesCount) + "/" + String.valueOf(cnt),
+                    uploadedFilesCount == cnt ? "success" : "failed");
+            log.info("upload end, cost {} ms", System.currentTimeMillis() - start);
+        });
     }
 
     private Path getTargetLocation(String filename, String savePath) {
@@ -72,10 +107,13 @@ public class StorageServiceImpl extends HttpServlet implements StorageService  {
                 dir = StorageDirConfig.DIR.IMAGES.toString().toLowerCase();
                 break;
             case "mov":
+            case "mp4":
+            case "mkv":
+            case "avi":
                 dir = StorageDirConfig.DIR.VIDEOS.toString().toLowerCase();
                 break;
             default:
-                dir = StorageDirConfig.DIR.DOCS.toString().toLowerCase();
+                dir = StorageDirConfig.DIR.DOCUMENTS.toString().toLowerCase();
                 break;
         }
         if (savePath != null && !savePath.isEmpty()) {
@@ -114,17 +152,17 @@ public class StorageServiceImpl extends HttpServlet implements StorageService  {
         return fileList;
     }
 
-    public String printDirectoryStructure(File dir) {
-        StringBuilder sb = new StringBuilder();
+    public String printDirectoryStructure(File dir, String type, boolean setType, StringBuilder sb) {
 
         if (dir != null && dir.isDirectory()) {
             sb.append("<ul>");
             for (File child : dir.listFiles()) {
+                if (setType) {
+                    type = child.getName();
+                }
                 if (child.isDirectory()) {
-                    sb.append("<li><a href=\"" + child.getName() + "\">" + child.getName() + "</a></li>\n");
-                    printDirectoryStructure(child);
-                } else {
-                    sb.append("<li>" + child.getName() + "</li>\n");
+                    sb.append("<li><a href=\"list?type=" + type + "&dir=" + child.getAbsolutePath().replace(path, "").replace("\\", "/") + "\">" + child.getName() + "</a></li>\n");
+                    printDirectoryStructure(child, type, false, sb);
                 }
             }
